# specs/grc-claw-openapi.yaml
openapi: 3.1.0
info:
  title: GRC_Claw API
  description: |
    GRC_Claw — open-source governance, risk, and compliance platform for agentic AI.
    
    This API provides four complementary interfaces:
    - **REST** — CRUD operations, search, and reporting
    - **GraphQL** — Complex nested queries for dashboards and analytics
    - **gRPC** — High-performance enforcement decisions and streaming
    - **Webhooks** — Event-driven integrations
    
    ## Authentication
    
    All API requests require authentication via one of:
    - **OAuth 2.1 / OIDC** — Bearer JWT token (RS256)
    - **API Key** — `Authorization: Bearer grc_live_...` or `grc_test_...`
    - **mTLS (SPIFFE)** — X.509 SVID for agent-to-PEP communication
    
    ## Rate Limiting
    
    Rate limit headers are included in all responses:
    - `X-RateLimit-Limit` — Request limit per window
    - `X-RateLimit-Remaining` — Remaining requests in current window
    - `X-RateLimit-Reset` — Unix timestamp when the window resets
    
    ## Error Format
    
    All errors follow RFC 7807 (Problem Details for HTTP APIs).
    
    ## Versioning
    
    Current version: `v1.0`. Version is specified in the URL path.
    Deprecation notices are sent via webhook 6 months before sunset.
  version: 1.0.0
  contact:
    name: GRC_Claw Architecture Team
    url: https://api.grc-claw.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.grc-claw.io
...
paths:
...
  /v1.0/agents:
    get:
      operationId: listAgents
      summary: List agents
      description: Returns paginated list of registered agents.
      tags: [Agents]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [model, agent, pipeline, endpoint]
        - name: framework
          in: query
          schema:
            type: string
            enum: [langchain, autogen, crewai, custom, mcp-server]
        - name: lifecycle_stage
          in: query
          schema:
            type: string
            enum: [proposed, approved, active, deprecated, terminated]
        - name: risk_tier
          in: query
          schema:
            type: string
            enum: [prohibited, high, limited, minimal]
        - name: trust_score_min
          in: query
          schema: { type: integer, minimum: 0, maximum: 100 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 500 }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      operationId: registerAgent
      summary: Register agent
      description: Registers a new agent in the registry.
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegisterRequest'
      responses:
        '201':
          description: Agent registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /v1.0/agents/{agent_id}:
    parameters:
      - name: agent_id
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getAgent
      summary: Get agent
      description: Returns full agent details including identity and trust score.
      tags: [Agents]
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      operationId: updateAgent
      summary: Update agent
      description: Updates an existing agent.
      tags: [Agents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentUpdateRequest'
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/agents/{agent_id}/trust-score:
    post:
      operationId: updateAgentTrustScore
      summary: Update agent trust score
      description: Updates the trust score for an agent.
      tags: [Agents]
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value, grade]
              properties:
                value:
                  type: integer
                  minimum: 0
                  maximum: 100
                grade:
                  type: string
                  enum: [A, B, C, D, F]
                reason:
                  type: string
      responses:
        '200':
          description: Trust score updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1.0/agents/{agent_id}/policy-bindings:
    post:
      operationId: bindPolicyToAgent
      summary: Bind policies to agent
      description: Binds one or more policies to an agent.
      tags: [Agents]
      parameters:
        - name: agent_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [policy_ids]
              properties:
                policy_ids:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Policies bound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: 'API key in format: Bearer grc_live_... or Bearer grc_test_...'
    mtlsAuth:
      type: mutualTLS
      description: mTLS with SPIFFE SVID
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail, code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        trace_id:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
    Pagination:
      type: object
      properties:
        next_cursor:
          type: string
        has_next:
          type: boolean
        total:
          type: integer
...
    Agent:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        type: { type: string }
        framework: { type: string }
        owner: { type: string }
        lifecycle_stage:
          type: string
          enum: [proposed, approved, active, deprecated, terminated]
        risk_tier:
          type: string
          enum: [prohibited, high, limited, minimal]
        capabilities:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              permissions:
                type: array
                items: { type: string }
              resource_scope: { type: string }
        identity:
          type: object
          properties:
            spiffe_id: { type: string }
            mtls_cert: { type: string }
            cert_expiry:
              type: string
              format: date-time
        trust_score:
          type: object
          properties:
            value: { type: integer }
            grade: { type: string }
            last_evaluated:
              type: string
              format: date-time
        policy_bindings:
          type: array
          items: { type: string }
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentRegisterRequest:
      type: object
      required: [name, type, framework, risk_tier]
      properties:
        name: { type: string }
        type: { type: string }
        framework: { type: string }
        owner: { type: string }
        risk_tier:
          type: string
          enum: [prohibited, high, limited, minimal]
        capabilities:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              permissions:
                type: array
                items: { type: string }
              resource_scope: { type: string }
    AgentUpdateRequest:
      type: object
      properties:
        name: { type: string }
        lifecycle_stage:
          type: string
          enum: [proposed, approved, active, deprecated, terminated]
        risk_tier:
          type: string
          enum: [prohibited, high, limited, minimal]
        capabilities:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              description: { type: string }
              permissions:
                type: array
                items: { type: string }
              resource_scope: { type: string }
    AgentListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        pagination:
          $ref: '#/components/schemas/Pagination'
...